+++ examples/gh-runner-gke-simple/main.tf
@@ -15,11 +15,14 @@
  */
 
 module "runner-gke" {
-  source         = "../../modules/gh-runner-gke"
-  create_network = true
-  project_id     = var.project_id
-  repo_name      = var.repo_name
-  repo_owner     = var.repo_owner
-  repo_url       = var.repo_url
-  gh_token       = var.gh_token
+  source  = "terraform-google-modules/github-actions-runners/google//modules/gh-runner-gke"
+  version = "~> 5.0"
+
+  project_id             = var.project_id
+  create_network         = true
+  cluster_suffix         = "k8s"
+  gh_app_id              = "123456"
+  gh_app_installation_id = "12345678"
+  gh_app_private_key     = "sample"
+  gh_config_url          = "https://github.com/ORGANIZATION"
 }