+++ examples/oidc-simple/main.tf
@@ -27,10 +27,13 @@ resource "google_project_iam_member" "project" {
 }
 
 module "oidc" {
-  source      = "../../modules/gh-oidc"
-  project_id  = var.project_id
-  pool_id     = "example-pool"
-  provider_id = "example-gh-provider"
+  source  = "terraform-google-modules/github-actions-runners/google//modules/gh-oidc"
+  version = "~> 5.0"
+
+  project_id          = var.project_id
+  pool_id             = "example-pool"
+  provider_id         = "example-gh-provider"
+  attribute_condition = "assertion.repository_owner_id=='1342004'"
   sa_mapping = {
     (google_service_account.sa.account_id) = {
       sa_name   = google_service_account.sa.name